+++ lib/utils.dart
@@ -68,4 +68,16 @@ class Utils {
     }
     return permStatus;
   }
+
+  static Future<bool> requestCameraPermission() async {
+    var status = await Permission.camera.request();
+    bool permStatus = false;
+    if (status.isGranted) {
+      print('Location Permission Granted');
+      permStatus = true;
+    } else if (status.isPermanentlyDenied) {
+      await openAppSettings();
+    }
+    return permStatus;
+  }
 }